Allow testing (shell) on linux/unix.
We rely upon /bin/echo, and the error-result of executing a
command that doesn't exist.
